payment-bank-api/pyproject.toml
[project]
name = "ufrpe-apsoo-card-bank-crud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
    "python-consul>=1.1.0",
]
